Вынес кнопку из меню в отдельный файл
